requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
trl>=0.7.0

# Data Processing
datasets>=2.15.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Training & Optimization
sentencepiece>=0.1.99
protobuf>=3.20.0
einops>=0.7.0
xformers>=0.0.23

# Evaluation & Metrics
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
sacrebleu>=2.3.0

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
omegaconf>=2.3.0

# Data Validation
jsonschema>=4.20.0
pydantic>=2.5.0

# Utilities
huggingface_hub>=0.19.0
safetensors>=0.4.0
python-json-logger>=2.0.7

# Development Tools
jupyter>=1.0.0
ipywidgets>=8.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# API & Serving (Optional)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
